Implement block evaluation
A block is a list of expressions.

There is an unusual result of evaluation that I discovered. With the
expression `["*", "x", "y"]` it is possible to have Eva evaluating the
expression `"x"` which would have caused an error. I had previously
been expecting that the simplest form a variable would have would be
inside a single element list i.e. `["x"]`. The result would be errors
stating "Unimplemented: "x"".

I expanded test coverage and `Eva.eval/2` to handle this case but I also
have a feeling that a fundamental assumption about the interpreter
expression forms is wrong. I can't identify the issue at this time.