Correction to variable evaluation tests
Going back over the course I realised I had an error in the variable
test assertions. The variable symbols `"x"` and `"y"` were not supposed
to be inside S-expressions (Elixir lists). The change I added to the
last commit already handles this case (so the interpreter is still
semantically correct).